+++ vector_quantization/__init__.py
@@ -1,3 +1,5 @@
+from rich.traceback import install
+
 from vector_quantization.finite_scalar_quantization import FSQ
 from vector_quantization.latent_quantization import LatentQuantize
 from vector_quantization.lookup_free_quantization import LFQ
@@ -8,6 +10,8 @@ from vector_quantization.residual_fsq import GroupedResidualFSQ, ResidualFSQ
 from vector_quantization.residual_lfq import GroupedResidualLFQ, ResidualLFQ
 from vector_quantization.residual_vq import GroupedResidualVQ, ResidualVQ
 from vector_quantization.vector_quantize_pytorch import VectorQuantize
+
+install(show_locals=False)
 
 __all__ = [
     "FSQ",