+++ Changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+* Fix warning regarding the deprecation of torch.cuda.amp.autocast.
+* Add the Rich traceback install in the '__init__' file of the package.
+* Update parameters and instanciation of the VQ module in the autoencoder example.
+
 ## [Version 1.17.0]
 
 * Create a new, meta Codebook class which can be used to create both EuclideanCodebook and CosineSimCodebook. Allows for further types of Codebooks.